# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Keywords API",
        "contact": {
            "name": "Yuriy",
            "email": "yurkaon@gmail.com"
        },
        "version": "1.0"
    },
    "paths": {
        "/api/register": {
            "post": {
                "tags": [
                    "register"
                ],
                "operationId": "/api/register",
                "parameters": [
                    {
                        "name": "register",
                        "in": "path",
                        "description": "Method for registretion",
                        "required": true,
                        "content": {
                            "application/json": {
                                "schema": {
                                    "required": [
                                        "name",
                                        "email",
                                        "password",
                                        "password_confirmation"
                                    ],
                                    "properties": {
                                        "name": {
                                            "type": "string",
                                            "format": "text",
                                            "example": "user1"
                                        },
                                        "email": {
                                            "type": "string",
                                            "format": "email",
                                            "example": "user1@mail.com"
                                        },
                                        "password": {
                                            "type": "string",
                                            "format": "password",
                                            "example": "PassWord12345"
                                        },
                                        "password_confirmation": {
                                            "type": "string",
                                            "format": "password",
                                            "example": "PassWord12345"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "User created successfully"
                    },
                    "409": {
                        "description": "Error: Cannot create user."
                    }
                }
            }
        },
        "/api/login": {
            "post": {
                "tags": [
                    "login"
                ],
                "operationId": "/api/login",
                "parameters": [
                    {
                        "name": "login",
                        "in": "path",
                        "description": "Method for log in",
                        "required": true,
                        "content": {
                            "application/json": {
                                "schema": {
                                    "required": [
                                        "email",
                                        "password"
                                    ],
                                    "properties": {
                                        "email": {
                                            "type": "string",
                                            "format": "email",
                                            "example": "user1@mail.com"
                                        },
                                        "password": {
                                            "type": "string",
                                            "format": "password",
                                            "example": "PassWord12345"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Logged in"
                    },
                    "409": {
                        "description": "Error: Unauthorized."
                    }
                }
            }
        },
        "/api/logout": {
            "post": {
                "tags": [
                    "logout"
                ],
                "operationId": "/api/logout",
                "parameters": [
                    {
                        "name": "logout",
                        "in": "path",
                        "description": "Method for log out",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Logged out"
                    }
                }
            }
        },
        "/api/categorys": {
            "post": {
                "tags": [
                    "Categorys"
                ],
                "operationId": "/api/categorys",
                "parameters": [
                    {
                        "name": "category",
                        "in": "path",
                        "description": "Controller using for creating new category",
                        "required": true,
                        "content": {
                            "application/json": {
                                "schema": {
                                    "required": [
                                        "category",
                                        "text",
                                        "images"
                                    ],
                                    "properties": {
                                        "category": {
                                            "type": "string",
                                            "format": "text",
                                            "example": "Apple"
                                        },
                                        "text": {
                                            "type": "string",
                                            "format": "text",
                                            "example": "Word1"
                                        },
                                        "images": {
                                            "type": "file",
                                            "format": "file",
                                            "example": "C:/images/123.png"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Category created",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "409": {
                        "description": "Error: Cannot create category."
                    }
                }
            }
        },
        "/api/images": {
            "get": {
                "tags": [
                    "Getting images using keyword"
                ],
                "operationId": "/api/images",
                "parameters": [
                    {
                        "name": "Getting images using keyword",
                        "in": "path",
                        "description": "Method for getting images using keyword",
                        "required": true,
                        "content": {
                            "application/json": {
                                "schema": {
                                    "required": [
                                        "keyword"
                                    ],
                                    "properties": {
                                        "keyword": {
                                            "type": "strind",
                                            "format": "text",
                                            "example": "Word1"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "DONE",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "409": {
                        "description": "Error: Cannot find image."
                    }
                }
            },
            "post": {
                "tags": [
                    "Upload image"
                ],
                "operationId": "/api/images",
                "parameters": [
                    {
                        "name": "upload image",
                        "in": "path",
                        "description": "Method for uploading images",
                        "required": true,
                        "content": {
                            "application/json": {
                                "schema": {
                                    "required": [
                                        "image"
                                    ],
                                    "properties": {
                                        "images": {
                                            "type": "file",
                                            "format": "file",
                                            "example": "C:/images/123.png"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "UPLOADED",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "409": {
                        "description": "Error: Cannot upload image."
                    }
                }
            }
        },
...
    },
...
}